fix: useProducts -> useProductlist
